Add routing, Remove unnecessary code

+++ backend/rp_backend/src/main/java/be/pxl/rp_backend/dto/UserDTO.java
@@ -6,7 +6,7 @@ public class UserDTO {
     private String first_name;
     private String last_name;
     private String email;
-    private long type;
+    private int type;
     private String role_name;
     private long pmi;
     private boolean use_pmi;
@@ -21,7 +21,7 @@ public class UserDTO {
     private Object phone_numbers;
     private String vanity_url;
     private String personal_meeting_url;
-    private long verified;
+    private int verified;
     private String pic_url;
     private String cms_user_id;
     private String account_id;
@@ -34,7 +34,7 @@ public class UserDTO {
     private String company;
     private String location;
     private Object custom_attributes;
-    private long login_type;
+    private int login_type;
     private String role_id;
     private String plan_united_type;
     private String manager;
@@ -75,11 +75,11 @@ public class UserDTO {
         this.email = email;
     }
 
-    public long getType() {
+    public int getType() {
         return type;
     }
 
-    public void setType(long type) {
+    public void setType(int type) {
         this.type = type;
     }
 
@@ -195,11 +195,11 @@ public class UserDTO {
         this.personal_meeting_url = personal_meeting_url;
     }
 
-    public long getVerified() {
+    public int getVerified() {
         return verified;
     }
 
-    public void setVerified(long verified) {
+    public void setVerified(int verified) {
         this.verified = verified;
     }
 
@@ -299,11 +299,11 @@ public class UserDTO {
         this.custom_attributes = custom_attributes;
     }
 
-    public long getLogin_type() {
+    public int getLogin_type() {
         return login_type;
     }
 
-    public void setLogin_type(long login_type) {
+    public void setLogin_type(int login_type) {
         this.login_type = login_type;
     }
